pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "sv-detection"
dynamic = ["version"]
description = "Automatic synaptic vesicle detection from FIB-SEM 3D volumes"
authors = [
    {name = "Neurobiology Lab", email = "neurobiology@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "scikit-image>=0.19.0",
    "pandas>=1.3.0",
    "pyarrow>=5.0.0",
    "zarr>=2.10.0",
    "dask[array]>=2021.9.0",
    "tqdm>=4.62.0",
    "numba>=0.56.0",
    "opencv-python>=4.5.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "napari>=0.4.15",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda11x>=10.0.0; platform_machine=='x86_64' and sys_platform=='linux'",
    "cupy-cuda12x>=12.0.0; platform_machine=='x86_64' and sys_platform=='linux'",
]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.9",
    "mypy>=0.900",
]

[project.scripts]
sv-detect = "sv_detection.cli:main"

[project.urls]
Homepage = "https://github.com/neurobiology-ut/synaptic-vesicle-detection"
Repository = "https://github.com/neurobiology-ut/synaptic-vesicle-detection"

[tool.setuptools_scm]
write_to = "sv_detection/_version.py"

[tool.setuptools.packages.find]
where = ["."]
include = ["sv_detection*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
